new: dev: Custom error pages
Error pages are served by the webserver when for some reason an error
occurred. This could be because of many things such as a typo,
unauthenticated users, or a backend server is not responding.

The goal of this way of building error pages here is to have a single
.html file output for every error page we want to serve. In this .html
there is minified CSS and webfonts included.

The error pages themselves are described using Markdown.

Right now only one error page is included, 404 - not-found.md.

Change-Id: I322bb65e879524e7b2de252d32afb6e843c2e60a
Signed-off-by: Christian Svensson <blue@cmd.nu>